+++ app/src/main/java/woowacourse/payments/MainActivity.kt
@@ -1,0 +1,20 @@
+package woowacourse.payments
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import woowacourse.payments.ui.components.PaymentCard
+import woowacourse.payments.ui.theme.AndroidpaymentsTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            AndroidpaymentsTheme {
+                PaymentCard()
+            }
+        }
+    }
+}
